fix: require a CSRF token for getting OAuth code; closes ON-4411
